# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sneak  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categorie": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Obtiener categorias",
                "description": "Obtener todos las categorias",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categorie"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/login-dealer": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login dealers",
                "description": "Login con email y contraseña para los repartidores",
                "operationId": "authLoginDealer",
                "requestBody": {
                    "description": "Ponga sus credenciales",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "info@sneak.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciales incorrectas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login-customer": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login customers",
                "description": "Login con email y contraseña para los clientes",
                "operationId": "authLoginCustomer",
                "requestBody": {
                    "description": "Ponga sus credenciales",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "josejuan@alcoyano.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciales incorrectas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Categorie": {
                "required": [
                    "id, name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Zapatillas"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Categorie"
                }
            },
...
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}